Create README.md
This project provides a "visitlist" for our animal adoption agency. Adopters can see the visitlist, delete animals from the list, then adopt one of the animals from their visitlist, most likely after they visit these animals at the shelter. In this project we create a class Visit, the instances of which would be visitclass objects. We also create an Animal class for animals in the shelter. We also use already created data_handling.py file to visualize our reports for the visit list